RAG/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.16
langchain-text-splitters>=0.0.1
openai>=1.12.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tiktoken>=0.5.2 